refactor(actions): move target user helper out of actions
